# Conflicts:
#	src/main/java/cn/rjtech/admin/vendorclass/VendorClassService.java

+++ src/main/java/cn/rjtech/admin/vendorclass/VendorClassService.java
@@ -15,7 +15,6 @@ import cn.rjtech.admin.cusfieldsmappingd.CusFieldsMappingDService;
 import cn.rjtech.admin.vendor.VendorService;
 import cn.rjtech.admin.vendoraddr.VendorAddrService;
 import cn.rjtech.enums.SourceEnum;
-import cn.rjtech.model.momdata.CustomerClass;
 import cn.rjtech.model.momdata.Vendor;
 import cn.rjtech.model.momdata.VendorAddr;
 import cn.rjtech.model.momdata.VendorClass;
@@ -50,6 +49,9 @@ public class VendorClassService extends BaseService<VendorClass> {
 
     private final VendorClass dao = new VendorClass().dao();
 
+    @Inject
+    private VendorService vendorService;
+
     @Override
     protected VendorClass dao() {
         return dao;
@@ -62,8 +64,6 @@ public class VendorClassService extends BaseService<VendorClass> {
 
     @Inject
     private CusFieldsMappingDService cusFieldsMappingdService;
-    @Inject
-    private VendorService            vendorService;
     @Inject
     private VendorAddrService        vendorAddrService;
 
@@ -175,41 +175,44 @@ public class VendorClassService extends BaseService<VendorClass> {
     }
 
     public Ret deleteVendorClassByIds(String ids) {
+        String[] split = ids.split(",");
+        String check = checkDelete(split);
+        if (StringUtils.isNotBlank(check)) {
+            return fail(check);
+        }
         boolean result = tx(() -> {
-            String[] split = ids.split(",");
             for (String id : split) {
                 //删除分类id
                 VendorClass vendorClass = findById(id);
                 if (vendorClass != null) {
-                    //删除分类id关联的子id
+                    //如果有下级分类，不允许删除
                     List<VendorClass> classListByIpid = findByIpid(id);
-                    classListByIpid.add(vendorClass);
+                    if (!classListByIpid.isEmpty()) {
+                        fail(vendorClass.getCVCName() + "：存在下级分类，不允许删除");
+                        return false;
+                    }
 
                     //删除分类编码关联的供应商档案Bd_Vendor
-                    for (VendorClass classByIpid : classListByIpid) {
-                        classByIpid.getIAutoId();//分类id
-                        classByIpid.getCVCCode();//供应商分类编码
-                        List<Vendor> vendors = vendorService
-                            .findByCVCCodeAndiVendorClassId(classByIpid.getCVCCode(), classByIpid.getIAutoId());
-                        for (Vendor vendor : vendors) {
-                            List<VendorAddr> vendorAddrs = vendorAddrService.findByIVendorId(vendor.getIAutoId());
-                            List<Long> vendorAddrscollect = vendorAddrs.stream().map(BaseVendorAddr::getIAutoId)
-                                .collect(Collectors.toList());
-                            List<String> strings = longConversionString(vendorAddrscollect);
-                            String addrIds = String.join(",", strings);
-                            //根据供应商档案删除Bd_VendorAddr
-                            vendorAddrService.deleteByIds(addrIds);
-                        }
-                        //删除供应商档案
-                        if (!vendors.isEmpty()) {
-                            List<Long> vendorcollect = vendors.stream().map(BaseVendor::getIAutoId).collect(Collectors.toList());
-                            List<String> vendorstrings = longConversionString(vendorcollect);
-                            String vendorids = String.join(",", vendorstrings);
-                            vendorService.deleteByIds(vendorids);
-                        }
-                        //删除分类id
-                        delete(classByIpid.getIAutoId());
+                    List<Vendor> vendors = vendorService
+                        .findByCVCCodeAndiVendorClassId(vendorClass.getCVCCode(), vendorClass.getIAutoId());
+                    for (Vendor vendor : vendors) {
+                        List<VendorAddr> vendorAddrs = vendorAddrService.findByIVendorId(vendor.getIAutoId());
+                        List<Long> vendorAddrscollect = vendorAddrs.stream().map(BaseVendorAddr::getIAutoId)
+                            .collect(Collectors.toList());
+                        List<String> strings = longConversionString(vendorAddrscollect);
+                        String addrIds = String.join(",", strings);
+                        //根据供应商档案删除Bd_VendorAddr
+                        vendorAddrService.deleteByIds(addrIds);
                     }
+                    //删除供应商档案
+                    if (!vendors.isEmpty()) {
+                        List<Long> vendorcollect = vendors.stream().map(BaseVendor::getIAutoId).collect(Collectors.toList());
+                        List<String> vendorstrings = longConversionString(vendorcollect);
+                        String vendorids = String.join(",", vendorstrings);
+                        vendorService.deleteByIds(vendorids);
+                    }
+                    //删除分类id
+                    delete(vendorClass.getIAutoId());
                 }
             }
             return true;
@@ -217,12 +220,27 @@ public class VendorClassService extends BaseService<VendorClass> {
         return ret(result);
     }
 
-    public List<String> longConversionString(List<Long> list){
-       return list.stream().map(x -> x + "").collect(Collectors.toList());
+    public List<String> longConversionString(List<Long> list) {
+        return list.stream().map(x -> x + "").collect(Collectors.toList());
     }
 
     public List<VendorClass> findByIpid(String ipid) {
         return find("SELECT * from Bd_VendorClass where iPid = ?", ipid);
+    }
+
+    public String checkDelete(String[] split) {
+        for (String id : split) {
+            //删除分类id
+            VendorClass vendorClass = findById(id);
+            if (vendorClass != null) {
+                //如果有下级分类，不允许删除
+                List<VendorClass> classListByIpid = findByIpid(id);
+                if (!classListByIpid.isEmpty()) {
+                    return vendorClass.getCVCName() + "：存在下级分类，不允许删除";
+                }
+            }
+        }
+        return null;
     }
 
     /**
@@ -318,6 +336,7 @@ public class VendorClassService extends BaseService<VendorClass> {
      * 树结构数据源
      */
     public List<JsTreeBean> getMgrTree(Long checkedId, int openLevel) {
+        List<Record> list = vendorService.List();
         List<VendorClass> supplierclasses = getMgrList();
         List<JsTreeBean> jsTreeBeanList = new ArrayList<>();
         for (VendorClass vendorClass : supplierclasses) {
@@ -328,6 +347,17 @@ public class VendorClassService extends BaseService<VendorClass> {
             }
             String text = "[" + vendorClass.getCVCCode() + "]" + vendorClass.getCVCName();
             String type = vendorClass.getCVCCode();
+            JsTreeBean jsTreeBean = new JsTreeBean(id, pid, text, type, "", false);
+            jsTreeBeanList.add(jsTreeBean);
+        }
+        for (Record record : list) {
+            Long id = record.get("iAutoId");
+            Object pid = record.get("iVendorClassId");
+            if (pid == null) {
+                pid = "#";
+            }
+            String text = "[" + record.get("cVenCode") + "]" + record.get("cVenName");
+            String type = record.get("cVenCode");
             JsTreeBean jsTreeBean = new JsTreeBean(id, pid, text, type, "", false);
             jsTreeBeanList.add(jsTreeBean);
         }
@@ -442,7 +472,8 @@ public class VendorClassService extends BaseService<VendorClass> {
         for (Record row : subList) {
             // 当前节点
             treeBeans.add(
-                new JsTreeBean(row.getStr("cvccode"), pid, row.getStr("cvcname") + "(" + row.getStr("cvccode") + ")", "default",
+                new JsTreeBean(row.getStr("cvccode"), pid, row.getStr("cvcname") + "(" + row.getStr("cvccode") + ")",
+                    "default",
                     null, true));
             // 追加子节点
             appendSubTree(treeBeans, row.getLong("iautoid"));
